--- a/pkg/storage/engine/bench_rocksdb_test.go
+++ b/pkg/storage/engine/bench_rocksdb_test.go
@@ -400,18 +400,23 @@ func BenchmarkMVCCGarbageCollect_RocksDB(b *testing.B) {
 	}
 }
 
-func BenchmarkBatchApplyBatchRepr(b *testing.B) {
+func BenchmarkBatchApplyBatchRepr_RocksDB(b *testing.B) {
 	if testing.Short() {
 		b.Skip("short flag")
 	}
 	ctx := context.Background()
-	for _, writeOnly := range []bool{false, true} {
-		b.Run(fmt.Sprintf("writeOnly=%t ", writeOnly), func(b *testing.B) {
-			for _, valueSize := range []int{10} {
-				b.Run(fmt.Sprintf("valueSize=%d", valueSize), func(b *testing.B) {
-					for _, batchSize := range []int{1000000} {
-						b.Run(fmt.Sprintf("batchSize=%d", batchSize), func(b *testing.B) {
-							runBatchApplyBatchRepr(ctx, b, setupMVCCInMemRocksDB, writeOnly, valueSize, batchSize)
+	for _, indexed := range []bool{false, true} {
+		b.Run(fmt.Sprintf("indexed=%t", indexed), func(b *testing.B) {
+			for _, sequential := range []bool{false, true} {
+				b.Run(fmt.Sprintf("seq=%t", sequential), func(b *testing.B) {
+					for _, valueSize := range []int{10} {
+						b.Run(fmt.Sprintf("valueSize=%d", valueSize), func(b *testing.B) {
+							for _, batchSize := range []int{10000} {
+								b.Run(fmt.Sprintf("batchSize=%d", batchSize), func(b *testing.B) {
+									runBatchApplyBatchRepr(ctx, b, setupMVCCInMemRocksDB,
+										indexed, sequential, valueSize, batchSize)
+								})
+							}
 						})
 					}
 				})