Add sequential vs random variant to this benchmark, and shrink the batch
size from 1,000,000 to 10,000. The 1m size batch was quite
unrealistic. Even 10k might be too large.

Old is RocksDB and new is Pebble. The reason for the discrepancy on the
`indexed=true/seq=true` variant is that RocksDB has a fast path for
in-order insertion, while Pebble does not (yet).

```
name                                                                         old time/op    new time/op     delta
BatchApplyBatchRepr/indexed=false/seq=false/valueSize=10/batchSize=10000-16     540µs ± 2%      187µs ± 2%   -65.41%  (p=0.000 n=10+9)
BatchApplyBatchRepr/indexed=false/seq=true/valueSize=10/batchSize=10000-16      538µs ± 2%      187µs ± 1%   -65.24%  (p=0.000 n=9+9)
BatchApplyBatchRepr/indexed=true/seq=false/valueSize=10/batchSize=10000-16     4.90ms ± 2%     2.08ms ± 3%   -57.66%  (p=0.000 n=10+10)
BatchApplyBatchRepr/indexed=true/seq=true/valueSize=10/batchSize=10000-16       954µs ± 1%     1400µs ± 1%   +46.80%  (p=0.000 n=10+10)
```

Release note: None